requirements.txt
streamlit
yfinance
pandas
plotly
google-generativeai
kaleido
openai
alpha_vantage
tiingo
torch
transformers
finnhub-python
sentencepiece
accelerate